Create descriptor pools when needed
Before if we ran out of set allocations in a pool, the program would
panic.
This change implements a dynamic pool management system, where descriptor
pools are stored along with the current allocations and the maximum
allocations allowed in the descriptor pool: when a descriptor set
needs to be created, a pool with enough free descriptor sets is searched
and used: if no one is found, a new pool is created.